backend/requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
torch
numpy
scikit-learn
